+++ packages/0x.js/src/utils/assert.ts
@@ -8,13 +8,13 @@ import { BigNumber } from '@0xproject/utils';
 import { Web3Wrapper } from '@0xproject/web3-wrapper';
 import * as _ from 'lodash';
 
-import { signatureUtils } from '../utils/signature_utils';
+import { isValidSignature } from '@0xproject/order-utils';
 
 export const assert = {
     ...sharedAssert,
     isValidSignature(orderHash: string, ecSignature: ECSignature, signerAddress: string) {
-        const isValidSignature = signatureUtils.isValidSignature(orderHash, ecSignature, signerAddress);
-        this.assert(isValidSignature, `Expected order with hash '${orderHash}' to have a valid signature`);
+        const isValid = isValidSignature(orderHash, ecSignature, signerAddress);
+        this.assert(isValid, `Expected order with hash '${orderHash}' to have a valid signature`);
     },
     async isSenderAddressAsync(
         variableName: string,
@@ -27,9 +27,5 @@ export const assert = {
             isSenderAddressAvailable,
             `Specified ${variableName} ${senderAddressHex} isn't available through the supplied web3 provider`,
         );
-    },
-    async isUserAddressAvailableAsync(web3Wrapper: Web3Wrapper): Promise<void> {
-        const availableAddresses = await web3Wrapper.getAvailableAddressesAsync();
-        this.assert(!_.isEmpty(availableAddresses), 'No addresses were available on the provided web3 provider');
     },
 };