+++ packages/contracts/util/types.ts
@@ -100,19 +100,3 @@ export enum ContractName {
     EtherDelta = 'EtherDelta',
     Arbitrage = 'Arbitrage',
 }
-
-export interface Artifact {
-    contract_name: ContractName;
-    networks: {
-        [networkId: number]: {
-            abi: ContractAbi;
-            solc_version: string;
-            keccak256: string;
-            optimizer_enabled: number;
-            unlinked_binary: string;
-            updated_at: number;
-            address: string;
-            constructor_args: string;
-        };
-    };
-}