+++ packages/deployer/src/index.ts
@@ -1,2 +1,3 @@
 export { Deployer } from './deployer';
 export { Compiler } from './compiler';
+export { ContractArtifact, ContractNetworks } from './utils/types';