+++ packages/order-utils/src/constants.ts
@@ -1,0 +1,3 @@
+export const constants = {
+    NULL_ADDRESS: '0x0000000000000000000000000000000000000000',
+};