+++ packages/sol-cov/src/utils.ts
@@ -4,6 +4,10 @@ export const utils = {
     compareLineColumn(lhs: LineColumn, rhs: LineColumn): number {
         return lhs.line !== rhs.line ? lhs.line - rhs.line : lhs.column - rhs.column;
     },
+    removeHexPrefix(hex: string): string {
+        const hexPrefix = '0x';
+        return hex.startsWith(hexPrefix) ? hex.slice(hexPrefix.length) : hex;
+    },
     isRangeInside(childRange: SingleFileSourceRange, parentRange: SingleFileSourceRange): boolean {
         return (
             utils.compareLineColumn(parentRange.start, childRange.start) <= 0 &&