+++ packages/web3-wrapper/CHANGELOG.md
@@ -5,11 +5,19 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v0.6.3 - _May 5, 2018_
+
+    * Dependencies updated
+
+## v0.6.2 - _May 4, 2018_
+
+    * Dependencies updated
+
 ## v0.6.1 - _April 18, 2018_
 
     * Dependencies updated
 
-## v0.6.0 - _April 12, 2018_
+## v0.6.0 - _April 11, 2018_
 
     * Make `isAddress` and `toWei` static (#501)
     * Add static methods `toUnitAmount` and `toBaseUnitAmount` (#501)
@@ -24,19 +32,19 @@ CHANGELOG
 
     * Rename `signTransactionAsync` to `signMessageAsync` for clarity (#465)
 
-## v0.3.0 - _March 18, 2018_
+## v0.3.0 - _March 17, 2018_
 
     * Add `web3Wrapper.takeSnapshotAsync`, `web3Wrapper.revertSnapshotAsync`, `web3Wrapper.mineBlockAsync`, `web3Wrapper.increaseTimeAsync` (#426)
     * Add `web3Wrapper.isZeroExWeb3Wrapper` for runtime instanceOf checks (#426)
     * Add a `getProvider` method (#444)
 
-## v0.2.0 - _March 4, 2018_
+## v0.2.0 - _March 3, 2018_
 
     * Ensure all returned user addresses are lowercase (#373)
     * Add `web3Wrapper.callAsync` (#413)
     * Make `web3Wrapper.estimateGas` accept whole `txData` instead of `data` (#413)
     * Remove `web3Wrapper.getContractInstance` (#413)
 
-## v0.1.12 - _February 9, 2018_
+## v0.1.12 - _February 8, 2018_
 
     * Fix publishing issue where .npmignore was not properly excluding undesired content (#389)